Small cleanup.
git-svn-id: https://svn.xiph.org/trunk/oggdsf@17192 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/lib/core/directshow/dsfOggDemux2/OggStreamMapper.h
@@ -29,7 +29,7 @@ public:
 
 	bool allStreamsReady();
 
-	unsigned long numPins()				{		return mPins.size();		}
+	size_t numPins()				{		return mPins.size();		}
 	OggDemuxOutputPin* getPinByIndex(unsigned long inIndex);
 
 protected:
